# Backend/src/swagger-spec.json
{"openapi":"3.0.0","info":{"title":"FitAI - AI-Powered Workout Planner API","version":"1.0.0","description":"\n# FitAI Backend API\n\nA comprehensive fitness tracking and AI-powered workout generation platform.\n\n## Features\n- 🤖 **AI Workout Generation** - Gemini 2.0 Flash powered intelligent workout creation\n- 💪 **Exercise Library** - 265+ exercises with detailed instructions\n- 📊 **Progress Tracking** - Track sets, reps, weight, and performance metrics\n- 🎯 **Goal Management** - Set and track fitness goals with AI recommendations\n- 📅 **Workout Templates** - Create and save reusable workout plans\n- 📈 **Analytics** - Comprehensive workout and progress analytics\n- 🔐 **Authentication** - JWT-based auth with OAuth (Google, GitHub)\n- 📸 **Media Upload** - Avatar and progress photos with S3 integration\n\n## Authentication\n\nMost endpoints require authentication using JWT tokens. Include the token in the Authorization header:\n\n```\nAuthorization: Bearer <your_token>\n```\n\nGet your token by registering or logging in via `/api/v1/auth/register` or `/api/v1/auth/login`.\n            ","contact":{"name":"FitAI API Support","email":"support@fitai.com"},"license":{"name":"MIT","url":"https://opensource.org/licenses/MIT"}},"servers":[{"url":"http://localhost:5000","description":"Development server"},{"url":"https://api.fitai.com","description":"Production server"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT","description":"JWT authorization header using the Bearer scheme. Example: \"Authorization: Bearer {token}\""}},"schemas":{"Error":{"type":"object","properties":{"success":{"type":"boolean","example":false},"message":{"type":"string","example":"Error message"},"statusCode":{"type":"integer","example":400},"errors":{"type":"array","items":{"type":"object","properties":{"field":{"type":"string"},"message":{"type":"string"}}}}}},"User":{"type":"object","properties":{"_id":{"type":"string","example":"507f1f77bcf86cd799439011"},"name":{"type":"string","example":"John Doe"},"email":{"type":"string","example":"john@example.com"},"profile":{"type":"object","properties":{"age":{"type":"number","example":25},"gender":{"type":"string","example":"male"},"height":{"type":"number","example":175},"weight":{"type":"number","example":70},"fitnessLevel":{"type":"string","example":"intermediate"}}}}},"Exercise":{"type":"object","properties":{"_id":{"type":"string","example":"507f1f77bcf86cd799439011"},"name":{"type":"string","example":"Barbell Bench Press"},"description":{"type":"string","example":"A compound upper body exercise targeting chest"},"primaryMuscleGroups":{"type":"array","items":{"type":"string"},"example":["chest","triceps"]},"difficulty":{"type":"string","enum":["beginner","intermediate","advanced"],"example":"intermediate"},"equipment":{"type":"array","items":{"type":"string"},"example":["barbell","bench"]}}},"Workout":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string","example":"Upper Body Strength"},"exercises":{"type":"array","items":{"type":"object","properties":{"exercise":{"type":"string","example":"507f1f77bcf86cd799439011"},"sets":{"type":"number","example":4},"reps":{"type":"number","example":10},"weight":{"type":"number","example":60}}}},"status":{"type":"string","enum":["planned","in_progress","completed"],"example":"planned"}}}}},"security":[{"bearerAuth":[]}],"tags":[{"name":"Authentication","description":"User authentication and registration"},{"name":"Users","description":"User profile management"},{"name":"Exercises","description":"Exercise library and search"},{"name":"Workouts","description":"Workout session management"},{"name":"AI","description":"AI-powered features (workout generation, exercise suggestions)"},{"name":"Templates","description":"Workout template CRUD operations"},{"name":"Goals","description":"Goal setting and tracking"},{"name":"Progress","description":"Progress tracking and history"},{"name":"Analytics","description":"Workout and progress analytics"},{"name":"Social","description":"Social features (sharing, feed)"}],"paths":{"/api/v1/ai/status":{"get":{"summary":"Check AI service status","tags":["AI"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"AI service status","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"available":{"type":"boolean","example":true},"model":{"type":"string","example":"gemini-2.0-flash"},"features":{"type":"array","items":{"type":"string"},"example":["workout_generation","exercise_suggestions","progress_analysis"]}}}}}}}}}}},"/api/v1/ai/workout-generate":{"post":{"summary":"Generate AI-powered workout plan","description":"Uses Gemini 2.0 Flash to create personalized workout based on user preferences","tags":["AI"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","properties":{"goal":{"type":"string","enum":["muscle_gain","weight_loss","strength","endurance","flexibility"],"example":"muscle_gain"},"duration":{"type":"number","description":"Workout duration in minutes","example":60},"difficulty":{"type":"string","enum":["beginner","intermediate","advanced"],"example":"intermediate"},"equipment":{"type":"array","items":{"type":"string"},"example":["barbell","dumbbells","bench"]},"targetMuscles":{"type":"array","items":{"type":"string"},"example":["chest","triceps","shoulders"]}}}}}},"responses":{"201":{"description":"Workout generated successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"workout":{"type":"object","properties":{"name":{"type":"string"},"exercises":{"type":"array","items":{"type":"object","properties":{"exercise":{"type":"string"},"sets":{"type":"number"},"reps":{"type":"number"},"restTime":{"type":"number"}}}}}},"aiMetadata":{"type":"object","properties":{"model":{"type":"string"},"generatedAt":{"type":"string"},"confidence":{"type":"number"}}}}}}}}}},"503":{"description":"AI service unavailable"}}}},"/api/v1/ai/exercise-suggest":{"post":{"summary":"Get AI-powered alternative exercise suggestions","description":"Suggests alternative exercises based on equipment, difficulty, or injury constraints","tags":["AI"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["exerciseId"],"properties":{"exerciseId":{"type":"string","description":"MongoDB ObjectId of the exercise","example":"507f1f77bcf86cd799439011"},"criteria":{"type":"object","properties":{"equipment":{"type":"array","items":{"type":"string"},"example":["bodyweight"]},"difficulty":{"type":"string","enum":["beginner","intermediate","advanced"],"example":"beginner"},"reason":{"type":"string","example":"No gym equipment available"}}}}}}}},"responses":{"200":{"description":"Alternative exercises suggested","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"alternatives":{"type":"array","items":{"type":"object","properties":{"_id":{"type":"string"},"name":{"type":"string"},"similarityScore":{"type":"number","example":85},"reason":{"type":"string","example":"Similar movement pattern with bodyweight"},"benefits":{"type":"string","example":"Can be performed anywhere without equipment"},"aiGenerated":{"type":"boolean"}}}}}}}}}}}}}},"/api/v1/ai/progress-insights":{"get":{"summary":"Get AI-powered progress analysis","description":"Analyzes workout history and provides insights and recommendations","tags":["AI"],"security":[{"bearerAuth":[]}],"responses":{"200":{"description":"Progress analysis completed","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"insights":{"type":"array","items":{"type":"string"}},"recommendations":{"type":"array","items":{"type":"string"}}}}}}}}}}}},"/api/v1/ai/goal-optimize":{"post":{"summary":"Get AI recommendations for goal optimization","tags":["AI"],"security":[{"bearerAuth":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["goalId"],"properties":{"goalId":{"type":"string"}}}}}},"responses":{"200":{"description":"Goal optimization suggestions"}}}},"/api/v1/auth/register":{"post":{"summary":"Register a new user","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["name","email","password"],"properties":{"name":{"type":"string","example":"John Doe"},"email":{"type":"string","format":"email","example":"john@example.com"},"password":{"type":"string","format":"password","minLength":8,"example":"Password123!"}}}}}},"responses":{"201":{"description":"User registered successfully","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"message":{"type":"string","example":"User registered successfully"},"data":{"type":"object","properties":{"id":{"type":"string"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}}}}}}},"400":{"description":"Validation error or user already exists","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Error"}}}}}}},"/api/v1/auth/login":{"post":{"summary":"Login with email and password","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email","password"],"properties":{"email":{"type":"string","format":"email","example":"john@example.com"},"password":{"type":"string","format":"password","example":"Password123!"}}}}}},"responses":{"200":{"description":"Login successful","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean"},"data":{"type":"object","properties":{"accessToken":{"type":"string"},"refreshToken":{"type":"string"}}}}}}}},"401":{"description":"Invalid credentials"},"404":{"description":"User not found"}}}},"/api/v1/auth/logout":{"post":{"summary":"Logout and invalidate refresh token","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["refreshToken"],"properties":{"refreshToken":{"type":"string"}}}}}},"responses":{"200":{"description":"Logged out successfully"}}}},"/api/v1/auth/refresh":{"post":{"summary":"Refresh access token using refresh token","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["refreshToken"],"properties":{"refreshToken":{"type":"string"}}}}}},"responses":{"200":{"description":"New access token generated","content":{"application/json":{"schema":{"type":"object","properties":{"accessToken":{"type":"string"}}}}}},"401":{"description":"Invalid or expired refresh token"}}}},"/api/v1/auth/forgot-password":{"post":{"summary":"Request password reset email","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["email"],"properties":{"email":{"type":"string","format":"email"}}}}}},"responses":{"200":{"description":"Password reset email sent"}}}},"/api/v1/auth/reset-password":{"post":{"summary":"Reset password using token from email","tags":["Authentication"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","required":["token","password"],"properties":{"token":{"type":"string"},"password":{"type":"string","minLength":8}}}}}},"responses":{"200":{"description":"Password reset successful"}}}},"/api/v1/auth/confirm-email":{"get":{"summary":"Confirm email address using token","tags":["Authentication"],"parameters":[{"in":"query","name":"token","required":true,"schema":{"type":"string"},"description":"Email confirmation token"}],"responses":{"200":{"description":"Email confirmed successfully"}}}},"/api/v1/exercises":{"get":{"summary":"Get all exercises with filtering and pagination","tags":["Exercises"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"query","name":"page","schema":{"type":"integer","default":1}},{"in":"query","name":"limit","schema":{"type":"integer","default":10}},{"in":"query","name":"muscleGroups","schema":{"type":"string"},"description":"Comma-separated muscle groups (e.g., chest,triceps)"},{"in":"query","name":"equipment","schema":{"type":"string"},"description":"Comma-separated equipment types"},{"in":"query","name":"difficulty","schema":{"type":"string","enum":["beginner","intermediate","advanced"]}},{"in":"query","name":"search","schema":{"type":"string"},"description":"Search in exercise name and description"}],"responses":{"200":{"description":"List of exercises (265 total in database)","content":{"application/json":{"schema":{"type":"object","properties":{"success":{"type":"boolean","example":true},"data":{"type":"array","items":{"$ref":"#/components/schemas/Exercise"}},"pagination":{"type":"object","properties":{"page":{"type":"integer"},"limit":{"type":"integer"},"total":{"type":"integer"},"totalPages":{"type":"integer"}}}}}}}}}}},"/api/v1/exercises/{id}":{"get":{"summary":"Get exercise by MongoDB ObjectId","tags":["Exercises"],"security":[{"bearerAuth":[]}],"parameters":[{"in":"path","name":"id","required":true,"schema":{"type":"string"},"description":"MongoDB ObjectId","example":"507f1f77bcf86cd799439011"}],"responses":{"200":{"description":"Exercise details"},"404":{"description":"Exercise not found"}}}}}}